+++ demo/src/components/UiSideNavContainer.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import startCase from 'lodash/startCase';
-import { Button, ListGroup, ListGroupItem, SidebarContainer, ScrollableContainer, SidebarHeader, Tooltip } from '../../../src/components';
+import { Button, ListGroup, ListGroupItem, Sidebar, ScrollableContainer, SidebarHeader, Tooltip } from '../../../src/components';
 
 import iconArrowLeft from '../assets/img/svg/icons/angle-left.svg';
 import UiComponents from '../config.module';
@@ -61,7 +61,7 @@ class UiSideNavContainer extends React.Component {
     const componentIds = Object.keys(UiComponents);
     const uiElementsList = this.generateElementListItem(componentIds);
     return (
-      <SidebarContainer id="sidenav" className="sidenav">
+      <Sidebar id="sidenav" className="sidenav">
         <SidebarHeader>
           <Button outline color="hover" className="btn-sidebar-close" onClick={this.toggleSideWrapper} id="btn-sidebar-close">
             <img src={iconArrowLeft} alt="<" />
@@ -74,7 +74,7 @@ class UiSideNavContainer extends React.Component {
           </ListGroup>
         </ScrollableContainer>
 
-      </SidebarContainer>
+      </Sidebar>
     );
   }
 }