+++ src/components/page-header/index.jsx
@@ -1,13 +1,12 @@
-// react
 import React from 'react';
 import PropTypes from 'prop-types';
 import './styles.scss';
 
 const PageHeader = ({ title, children }) => (
-  <div className="page-header">
+  <header className="page-header">
     {title && <h1 className="page-header-title">{title}</h1>}
     {children}
-  </div>
+  </header>
 );
 
 PageHeader.propTypes = {