+++ src/components/sidebar-header/index.jsx
@@ -1,22 +1,21 @@
-// react
 import React from 'react';
 import PropTypes from 'prop-types';
 import './styles.scss';
 
 
-const SideBarHeader = ({ title, children }) => (
+const SidebarHeader = ({ title, children }) => (
   <div className="sidebar-header" >
     <h3 className="sidebar-header-title">{title}</h3>{children}
   </div>
 );
 
-SideBarHeader.propTypes = {
+SidebarHeader.propTypes = {
   title: PropTypes.string,
   children: PropTypes.node,
 };
 
-SideBarHeader.defaultProps = {
+SidebarHeader.defaultProps = {
   title: '',
 };
 
-export default SideBarHeader;
+export default SidebarHeader;